Associations (#1)
This PR adds initial support for associations.

Associations allow for automatically creating another entity when one is
persisted. This can be used, for instance, to construct parent entities
in a hierarchy.

There's still some rough edges in the API that I'd like to get ironed
out, but right now it is functional.